backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Bonfix Ngetich"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
djangorestframework-jwt = "^1.11.0"
django-cors-headers = "^3.7.0"
cookiecutter = "^1.7.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = {version = "^21.6b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
